requirements.txt
sentence-transformers==3.2.0
psycopg2-binary==2.9.9
pymongo==4.10.1
tqdm==4.66.5
python-dotenv==1.0.1
